+++ tailwind-prefixer.js
@@ -1,4 +1,4 @@
-module.exports = function (content, prefix, sizes = ['sm', 'md', 'lg', 'xl', '2xl']) {
+module.exports = function (content, prefix = '', sizes = ['sm', 'md', 'lg', 'xl', '2xl']) {
   if (content.match('^// prefix')) {
     content = content.replace(/(['`])(.*)\1/g, function(match) {
       return `${match.match(/(['`])/)[1]}${match.match(/(['`])(.*)\1/)[2].split(' ').map((c) => {