+++ themes/vueform/templates/ElementLabel.vue
@@ -12,11 +12,27 @@
           container_sm: 'vf-label-sm',
           container_md: '',
           container_lg: 'vf-label-lg',
+          container_horizontal_sm: 'vf-horizontal-label-sm',
+          container_horizontal_md: 'vf-horizontal-label',
+          container_horizontal_lg: 'vf-horizontal-label-lg',
+          container_vertical_sm: 'vf-vertical-label-sm',
+          container_vertical_md: 'vf-vertical-label',
+          container_vertical_lg: 'vf-vertical-label-lg',
           wrapper: '',
           $container: (classes, { el$, Size }) => ([
             classes.container,
             classes[`container_${Size}`],
             !el$.inline ? el$.columnsClasses.label : null,
+            el$.cols.default.label < 12 ? classes[`container_horizontal_${Size}`] : classes[`container_vertical_${Size}`],
+            ...(Object.keys(el$.cols).length > 1 ? (el$.$vueform.config.breakpoints||['sm', 'md', 'lg', 'xl', '2xl']).map((breakpoint) => {
+              if (!el$.cols[breakpoint]?.label) {
+                return null
+              }
+
+              return el$.cols[breakpoint].label < 12
+                ? classes[`container_horizontal_${Size}`].map(c => `vf-${breakpoint}:${c}`).join(' ')
+                : classes[`container_vertical_${Size}`].map(c => `vf-${breakpoint}:${c}`).join(' ')
+            }) : []),
           ]),
         }
       }
@@ -25,49 +41,5 @@
 </script>
 
 <style lang="scss">
-  .vf-label {
-    margin-bottom: 0;
-    display: flex;
-    align-items: flex-start;
-    padding-right: var(--vf-gutter);
-    padding-bottom: calc(var(--vf-gutter) / 3);
-    font-size: var(--vf-font-size);
-    line-height: var(--vf-line-height);
-    letter-spacing: var(--vf-letter-spacing);
-
-    &.vf-label-sm {
-      padding-right: var(--vf-gutter-sm);
-      padding-bottom: calc(var(--vf-gutter-sm) / 3);
-      font-size: var(--vf-font-size-sm);
-      line-height: var(--vf-line-height-sm);
-      letter-spacing: var(--vf-letter-spacing-sm);
-    }
-
-    &.vf-label-lg {
-      padding-right: var(--vf-gutter-lg);
-      padding-bottom: calc(var(--vf-gutter-lg) / 3);
-      font-size: var(--vf-font-size-lg);
-      line-height: var(--vf-line-height-lg);
-      letter-spacing: var(--vf-letter-spacing-lg);
-    }
-  }
-
-  .vf-form-horizontal {
-    .vf-text-type {
-      .vf-label {
-        padding-top: calc(var(--vf-border-width-input-t) + var(--vf-py-input));
-        padding-bottom: 0;
-
-        &.vf-label-sm {
-          padding-top: calc(var(--vf-border-width-input-t) + var(--vf-py-input-sm));
-          padding-bottom: 0;
-        }
-
-        &.vf-label-lg {
-          padding-top: calc(var(--vf-border-width-input-t) + var(--vf-py-input-lg));
-          padding-bottom: 0;
-        }
-      }
-    }
-  }
+  // Styles can be found at scss/_label.scss
 </style>